+++ commands/clone.js
@@ -1,5 +1,5 @@
 const chalk = require("chalk");
-const { askForProjectName } = require("../lib/app-name");
+const { askForProjectName, getAppName } = require("../lib/app-name");
 const { askForCorrectRepoNames } = require("../lib/clone");
 const linkValetSite = require("../tasks/valet");
 const installDependencies = require("../tasks/dependencies");
@@ -8,8 +8,10 @@ const util = require("util");
 const { getTrellisPath } = require("../lib/trellis");
 const prompts = require("prompts");
 const exec = util.promisify(require("child_process").exec);
+const { spawnSync } = require("child_process");
 const fs = require("fs");
 const dbImport = require("./db");
+const { getAdminUrl } = require("../lib/wordpress");
 
 async function cloneLisaProject() {
   console.log(
@@ -75,9 +77,13 @@ async function cloneLisaProject() {
     cwd: trellisPath,
   });
 
+  console.log(chalk.green(`🎉 Generated .env file with trellis-cli!.`));
+
   await exec(`wp db create`, {
     cwd: `${apiName}/site`,
   });
+
+  console.log(chalk.green(`🎉 Created empty local database!.`));
 
   let { doDbImport } = await prompts([
     {
@@ -90,6 +96,39 @@ async function cloneLisaProject() {
   if (doDbImport) {
     await dbImport();
   }
+
+  spawnSync(`vercel link`, [], {
+    cwd: appName,
+    stdio: "inherit",
+    shell: true,
+  });
+
+  spawnSync(`vercel env pull .env.local`, [], {
+    cwd: appName,
+    stdio: "inherit",
+    shell: true,
+  });
+
+  console.log(
+    chalk.greenBright(
+      `🎉 Generated .env.local file with enviroments variables from Vercel.`
+    )
+  );
+
+  let adminUrl = await getAdminUrl();
+
+  console.log();
+  console.log(chalk.bold.greenBright(`🎉🎉🎉 All done! 🎉🎉🎉`));
+  console.log(
+    chalk.green(`Admin URL: ${chalk.underline(adminUrl)}/wp/wp-admin`)
+  );
+  console.log(
+    chalk.green(
+      `Run this command for local development: ${chalk.underline(
+        `cd ${appName} && yarn dev`
+      )}`
+    )
+  );
 }
 
 module.exports = cloneLisaProject;